fix: change default cache location
The default location for the prettier cache is within the `node_modules`
folder.  Whilst this is typically in a repository `.gitignore` file
for projects that use `node` for those that do not this can lead to the
accidental committal of unneccessary content (e.g. via `git add .`).

Prettier will delete the cache file if not being executed with `--cache`
but only if that `--cache-location` is different to the default.

Ref: https://github.com/prettier/prettier/issues/13032

Signed-off-by: Adam Moss <2951486+adam-moss@users.noreply.github.com>